added product and vendor model mapping

+++ graphql/resolvers/product.js
@@ -1,9 +1,10 @@
 const Product = require('../../models/product');
+const Vendor = require('../../models/vendor');
 
 exports.queryResolver = {
   getAllProducts: async (_, args) => {
     try {
-      return await Product.find();
+      return await Product.find().populate('vendors');
     } catch (err) {
       throw err;
     }
@@ -11,7 +12,8 @@ exports.queryResolver = {
   getProductByName: async (_, args) => {
     try {
       const { name } = args;
-      return await Product.findOne({ name: name });
+      const result = await Product.findOne({ name: name });
+      return await result.populate('vendors').execPopulate();
     } catch (err) {
       throw err;
     }
@@ -21,13 +23,19 @@ exports.queryResolver = {
 exports.mutationResolver = {
   addProduct: async (_, args) => {
     try {
-      let { name, category } = args.productInput;
+      let { name, category, vendorID } = args.productInput;
       const product = new Product({
         name: name,
         category: category,
+        vendors: vendorID,
       });
       const result = await product.save();
-      return result;
+      for (const item of vendorID) {
+        const vendor = await Vendor.findOne({ _id: item });
+        vendor.products.push(product._id);
+        vendor.save();
+      }
+      return await result.populate('vendors').execPopulate();
     } catch (err) {
       throw err;
     }
@@ -35,6 +43,19 @@ exports.mutationResolver = {
   removeProduct: async (_, args) => {
     try {
       let { name, category } = args.productInput;
+      const product = await Product.findOne({
+        name: name,
+        category: category,
+      });
+
+      if (!product) return false;
+      const { _id, vendors } = product;
+      for (const item of vendors) {
+        const vendor = await Vendor.findById(item);
+        const index = vendor.products.indexOf(_id);
+        vendor.products.splice(index, 1);
+        vendor.save();
+      }
       const result = await Product.deleteOne({
         name: name,
         category: category,
